+++ addOns/automation/CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 - Maintenance changes.
 - Rely on spider add-on (Issue 3113).
+- passiveScan-config job resets the state, as scanOnlyInScope is often confusing in the GUI. 
 
 ### Fixed
 - Correct loading of custom scripts (e.g. Zest).